pyproject.toml
[tool.poetry]
name = "status-villain"
version = "0.0.0"
description = "The CLI Standup bot for Engineers"
authors = ["Bastien Boutonnet <bastien.b1@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
SQLAlchemy = "^1.4.22"
psycopg2 = "^2.9.1"
questionary = "^1.10.0"
pydantic = "^1.8.2"
rich = "^10.7.0"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
pytest-cov = "^2.12.1"
pre-commit = "^2.13.0"
ipython = "^7.26.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
status-villain="status_villain.main:main"
